Broken HTML
Hi, guys,

It's surely that something is missing. I have guessed and placed the missing closing tag and blue bottom buttons returned to the right position. May be it's not the source of the issue but I bet you know better. In case you have no better solution use this one because the way it is now is not good at all.